Fold generation defaults into smart field service
